requirements.txt
pandas==2.1.4
Pillow==10.2.0
streamlit==1.30.0